requirements.txt
numpy
matplotlib
spyder
ffmpeg-python
imageio-ffmpeg
imageio